+++ components/DetailPageHighlightsSimple/DetailPageHighlightsSimple.js
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import { Notification } from '../../index';
+import { Notification } from '../carbon';
 
 const propTypes = {
   artifact: PropTypes.object,
