+++ components/DetailPageSidebar/DetailPageSidebar.js
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import classnames from 'classnames';
-import { Tag } from '../../index';
+import { Tag } from '../carbon';
 
 class DetailPageSidebar extends Component {
   static propTypes = {
