+++ components/carbon.js
@@ -1,0 +1,4 @@
+import Notification from '../node_modules/carbon-components-react/cjs/components/Notification';
+import Tag from '../node_modules/carbon-components-react/cjs/components/Tag';
+
+export { Notification, Tag };
