+++ index.js
@@ -1,11 +1,3 @@
-import Notification from 'carbon-components-react/cjs/components/Notification';
-import Tag from 'carbon-components-react/cjs/components/Tag';
-
-export {
-  Notification,
-  Tag,
-};
-
 export DetailPageSidebar from './components/DetailPageSidebar/DetailPageSidebar.js';
 export DetailPageHighlights from './components/DetailPageHighlights/DetailPageHighlights.js';
 export DetailPageHighlightsSimple from './components/DetailPageHighlightsSimple/DetailPageHighlightsSimple.js';
